requirements.txt
Django==2.2.3
Pillow==6.1.0
boto3==1.9.205
dj-database-url==0.5.0
django-ckeditor==5.7.1
django-storages==1.7.1
djangorestframework==3.10.2
gunicorn==19.9.0
psycopg2==2.8.3
whitenoise==4.1.2
